requirements.txt
absl-py==1.4.0
addict==2.4.0
apex==0.1
astunparse==1.6.3
batchgenerators==0.25
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
einops==0.6.0
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.2
future==0.18.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.2
idna==3.4
imageio==2.26.0
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.0
linecache2==1.0.0
lit==15.0.7
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mmcv==1.7.1
monai==1.1.0
mpmath==1.2.1
networkx==3.0
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Pillow==9.3.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
SimpleITK==2.2.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.6.12
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
traceback2==1.4.0
triton==2.0.0
typing_extensions==4.4.0
unittest2==1.1.0
urllib3==1.26.13
Werkzeug==2.2.3
wrapt==1.14.1
yapf==0.32.0
